+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
+## HardPy 0.7.0
+
 * Add a marker **attempt** to indicate the number of attempts to run a test before it passes successfully.
 * Add a method **get_current_attempt** to get the number of the current attempt.
 * Add the ability to run multiple dialog boxes in one test.